Refactored names to align with functionality of class/fragment
