+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".KeepingUpWithSchoolActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
