+++ app/src/main/java/edu/uc/clemenks/keepingupwithschool/KeepingUpWithSchoolActivity.kt
@@ -1,0 +1,20 @@
+package edu.uc.clemenks.keepingupwithschool
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import edu.uc.clemenks.keepingupwithschool.ui.main.ImageFragment
+
+class KeepingUpWithSchoolActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.main_activity)
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                .replace(R.id.container, ImageFragment.newInstance())
+                .commitNow()
+        }
+    }
+
+
+}
