--- app/src/main/java/edu/uc/clemenks/keepingupwithschool/MainActivity.kt
@@ -1,30 +1,0 @@
-package edu.uc.clemenks.keepingupwithschool
-
-import android.app.Activity
-import android.content.Intent
-import android.graphics.Bitmap
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.provider.MediaStore
-import android.widget.ImageView
-import android.widget.Toast
-import edu.uc.clemenks.keepingupwithschool.ui.main.MainFragment
-import kotlinx.android.synthetic.main.main_activity.*
-import kotlinx.android.synthetic.main.main_fragment.*
-
-class MainActivity : AppCompatActivity() {
-
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.main_activity)
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, MainFragment.newInstance())
-                .commitNow()
-        }
-    }
-
-
-}
