+++ app/src/main/res/layout/main_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainFragment">
+    tools:context=".ui.main.ImageFragment">
 
     <TextView
         android:id="@+id/lblKeepUWS"
@@ -61,6 +61,46 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <TextView
+        android:id="@+id/lblLatitide"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/latitude"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/lblKeepUWS" />
+
+    <TextView
+        android:id="@+id/lblLongitude"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/longitude"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/lblLatitide" />
+
+    <TextView
+        android:id="@+id/lblLongitudeValue"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/zeropointzero"
+        app:layout_constraintStart_toEndOf="@+id/lblLongitude"
+        app:layout_constraintTop_toBottomOf="@+id/lblLatitudeValue" />
+
+    <TextView
+        android:id="@+id/lblLatitudeValue"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:text="@string/zeropointzero"
+        app:layout_constraintStart_toEndOf="@+id/lblLatitide"
+        app:layout_constraintTop_toBottomOf="@+id/lblKeepUWS" />
+
     <LinearLayout
         android:id="@+id/linearLayout"
         android:layout_width="match_parent"
@@ -99,43 +139,5 @@
         app:layout_constraintTop_toBottomOf="@+id/linearLayout"
         tools:src="@tools:sample/avatars" />
 
-    <TextView
-        android:id="@+id/lblLatitide"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:text="@string/latitude"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/lblKeepUWS" />
 
-    <TextView
-        android:id="@+id/lblLongitude"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:text="@string/longitude"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/lblLatitide" />
-
-    <TextView
-        android:id="@+id/lblLatitudeValue"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:text="@string/zeropointzero"
-        app:layout_constraintStart_toEndOf="@+id/lblLatitide"
-        app:layout_constraintTop_toBottomOf="@+id/lblKeepUWS" />
-
-    <TextView
-        android:id="@+id/lblLongitudeValue"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:text="@string/zeropointzero"
-        app:layout_constraintStart_toEndOf="@+id/lblLongitude"
-        app:layout_constraintTop_toBottomOf="@+id/lblLatitudeValue" />
 </androidx.constraintlayout.widget.ConstraintLayout>
